renamed:    h2o_portal/h2o_portal/H2O_JS.js -> Storing old/H2O_JS.js 	renamed:    h2o_portal/H2O_Social_Tool.html -> Storing old/H2O_Social_Tool.html 	renamed:    h2o_portal/h2o_portal/facebooklogo.png -> Storing old/facebooklogo.png 	renamed:    h2o_portal/h2o_portal/groupme.jpg -> Storing old/groupme.jpg 	renamed:    h2o_portal/h2o_portal/h2oLogo.png -> Storing old/h2oLogo.png 	renamed:    h2o_portal/h2oStyles.css -> Storing old/h2oStyles.css 	renamed:    h2o_models.py -> Storing old/h2o_models.py 	renamed:    h2o_portal/index.html -> Storing old/index.html 	renamed:    h2o_portal/index.js -> Storing old/index.js 	renamed:    h2o_portal/h2o_portal/reset.css -> Storing old/reset.css 	new file:   h2o_portal/__pycache__/h2o_models.cpython-37.pyc 	new file:   h2o_portal/db.sqlite3 	new file:   h2o_portal/h2o_portal/__pycache__/__init__.cpython-37.pyc 	new file:   h2o_portal/h2o_portal/__pycache__/models.cpython-37.pyc 	new file:   h2o_portal/h2o_portal/__pycache__/serializers.cpython-37.pyc 	new file:   h2o_portal/h2o_portal/__pycache__/settings.cpython-37.pyc 	new file:   h2o_portal/h2o_portal/__pycache__/urls.cpython-37.pyc 	new file:   h2o_portal/h2o_portal/__pycache__/views.cpython-37.pyc 	new file:   h2o_portal/h2o_portal/__pycache__/wsgi.cpython-37.pyc 	new file:   h2o_portal/h2o_portal/migrations/0002_auto_20191107_1949.py 	new file:   h2o_portal/h2o_portal/migrations/0003_remove_post_picture.py 	new file:   h2o_portal/h2o_portal/migrations/__pycache__/0001_initial.cpython-37.pyc 	new file:   h2o_portal/h2o_portal/migrations/__pycache__/0002_auto_20191107_1949.cpython-37.pyc 	new file:   h2o_portal/h2o_portal/migrations/__pycache__/0003_remove_post_picture.cpython-37.pyc 	new file:   h2o_portal/h2o_portal/migrations/__pycache__/__init__.cpython-37.pyc 	modified:   h2o_portal/h2o_portal/models.py 	modified:   h2o_portal/h2o_portal/serializers.py 	modified:   h2o_portal/h2o_portal/urls.py 	modified:   h2o_portal/h2o_portal/views.py 	new file:   h2o_portal/urls.py 	renamed:    h2o_portal/manage.py -> manage.py
