+++ Storing old/h2oLogo.png
