--- h2o_portal/H2O_Social_Tool.html
@@ -1,115 +1,0 @@
-<!doctype html>
-<html lang="en-US">
-    
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>H2O Social Media Tool</title>
-  <link rel="stylesheet" href="h2o_portal/reset.css" type="text/css">
-  <link rel="stylesheet" href="h2oStyles.css" type="text/css">
-</head>
-<img id="h2opic" src="h2o_portal/h2oLogo.png" alt="H2O Logo" width="360" height="187">
-<header>H2O Social Media Tool</header>
-
-<body>
-  <h1>Make a Selection</h1>
-  <div id="options1">
-    <ul id="selections">
-      <li><button class="button" onclick="RevealWindow('CredDisp')">Manage Credentials</button></li>
-      <li><button class="button" onclick="RevealWindow('CreatePostDisp')">Create Post</button></li>
-      <li><button class="button" onclick="RevealWindow('ListSchPostsDisp')">List Scheduled Posts</button></li>
-      <li><button class="button" onclick="RevealWindow('DelSchPostsDisp')">Delete Scheduled Posts</button></li>
-      <li><button class="button" onclick="RevealWindow('CompPostsDisp')">Completed Posts</button></li>
-    </ul>
-  </div>
-
-  <div class="appauth">
-    <button class="button" onclick="RevealWindow('AppAuth')">
-      <h2>Login / Signup</h2>
-      <div id="AppAuth" class="revealpane"> 
-          <p>test login</p>
-      </div>
-    </button>
-  </div>
-
-  <div id="CredDisp" class="revealpane">     
-    <button class="SelectionButton" onclick="RevealCheckBox('RevealFacebookLogin')">Facebook</button>
-      <div id="RevealFacebookLogin" class="revealpaneinner">
-        <form action="" method="post">
-          <div class="imgcontainer">
-            <img src="h2o_portal/facebooklogo.png" alt="Facebooklogo" class="smallpic">
-          </div>
-          <div class="container">
-            <label for="uname"><b>Username</b></label>
-            <input type="text" placeholder="Enter Username" name="uname" required>
-            <label for="psw"><b>Password</b></label>
-            <input type="password" placeholder="Enter Password" name="psw" required>
-            <button type="submit" class="submitbutton">Store Credentials</button>
-          </div>
-        </form> 
-      </div>
-
-    <button class="SelectionButton" onclick="RevealCheckBox('RevealInstagramLogin')">Instagram</button>
-      <div id="RevealInstagramLogin" class="revealpaneinner">
-        <form action="" method="post">
-          <div class="imgcontainer">
-            <img src="h2o_portal/InstagramLogo.png" alt="InstagramLogo" class="smallpic">
-          </div>
-          <div class="container">
-            <label for="uname"><b>Username</b></label>
-            <input type="text" placeholder="Enter Username" name="uname" required>
-            <label for="psw"><b>Password</b></label>
-            <input type="password" placeholder="Enter Password" name="psw" required>
-            <button type="submit" class="submitbutton">Store Credentials</button>
-          </div>
-        </form>
-      </div>
-    
-    <button class="SelectionButton" onclick="RevealCheckBox('RevealGroupMeLogin')">GroupMe</button>
-      <div id="RevealGroupMeLogin" class="revealpaneinner">
-        <form action="" method="post">
-          <div class="imgcontainer">
-            <img src="h2o_portal/groupme.jpg" alt="GroupMeLogo" class="smallpic">
-          </div>
-          <div class="container">
-            <label for="uname"><b>Username</b></label>
-            <input type="text" placeholder="Enter Username" name="uname" required>
-            <label for="psw"><b>Password</b></label>
-            <input type="password" placeholder="Enter Password" name="psw" required>
-            <button type="submit" class="submitbutton">Store Credentials</button>
-          </div>
-        </form>
-      </div>
-    </div>
-  </div> <!--Closes Credential Pane-->
-
-  <div id="CreatePostDisp" class="revealpane">
-    <p>Create Post Pane</p>
-    <p>Select Platform Destination (Facebook, Instagram, GroupMe, or All)</p>
-    <p>Select Picture</p>
-    <p>Input Text:</p>
-  </div> <!--Closes Create Post Pane-->
-
-  <div id="ListSchPostsDisp" class="revealpane">
-    <p>List Scheduled Pane</p>
-    <p>Display Top 10 posts</p>
-    <p>Next 10 posts</p>
-  </div> <!--Closes List Scheduled Pane-->
-
-  <div id="DelSchPostsDisp" class="revealpane">
-    <p>Delete Scheduled Pane</p>
-    <p>Display Same as List Posts but with Delete Button avaiable</p>
-    <p>Must have admin permissions for other people's posts</p> 
-    <p>Your own posts should only be displayed if you are not an admin</p>
-    <p>Have a delete other posts button and require password again before that panel</p>
-  </div> <!--Closes Delete Scheduled Pane-->
-
-  <div id="CompPostsDisp" class="revealpane">
-    <p>Completed Pane</p>
-    <p>Once post is completed display latest 10</p>
-    <p>Next 10 posts</p>
-  </div> <!--Closes Completed Pane-->
-</body>
-
-<script src="h2o_portal/H2O_JS.js" type="text/javascript"></script>
-</html>
