+++ h2o_portal/h2o_portal/serializers.py
@@ -1,4 +1,3 @@
-from django.contrib.auth.models import User, Group
 from rest_framework import serializers
 from h2o_models import Credentials, Post
