+++ h2o_portal/h2o_portal/views.py
@@ -2,12 +2,15 @@ from rest_framework import generics
 from h2o_portal.models import Post, Credentials
 from h2o_portal.serializers import PostSerializer, CredentialSerializer
 
+
 class PostViewSet(generics.ListCreateAPIView):
-    serializer_class = PostSerializer
     queryset = Post.objects.all()
+    serializer_class = PostSerializer
+    
 
 
 class CredentialViewSet(generics.ListCreateAPIView):
-    serializer_class = CredentialSerializer
     queryset = Credentials.objects.all()
+    serializer_class = CredentialSerializer
+
